+++ odoo/local-src/specific_purchase/models/__init__.py
@@ -8,3 +8,4 @@ from . import bank_holiday
 from . import purchase_config_settings
 from . import purchase_order_cancel
 from . import product_supplierinfo
+from . import procurement_order