+++ odoo/songs/install/data_full.py
@@ -41,6 +41,16 @@ def import_suppliers(ctx):
         load_csv_stream(ctx, Partner, content, delimiter=',')
 
 
+def import_additional_suppliers_data(ctx):
+    """ Importing additional supplier data from csv """
+
+    load_ctx = ctx.env.context.copy()
+    load_ctx.update({'tracking_disable': True})
+    Partner = ctx.env['res.partner'].with_context(load_ctx)
+    for content in get_files(req, 'data/install/supplier_add_data.csv.csv'):
+        load_csv_stream(ctx, Partner, content, delimiter=',')
+
+
 @anthem.log
 def import_clients(ctx):
     """ Importing clients from csv"""